+++ matdb/calculators/basic.py
@@ -78,7 +78,7 @@ class AsyncCalculator(object):
         pass
 
     @abc.abstractmethod
-    def to_dict(self, folder):
+    def to_dict(self):
         """Writes the current version number of the code being run to a
         dictionary along with the parameters of the code.
 
@@ -136,7 +136,7 @@ class SyncCalculator(object):
         pass
 
     @abc.abstractmethod
-    def to_dict(self, folder):
+    def to_dict(self):
         """Writes the current version number of the code being run to a
         dictionary along with the parameters of the code.